refactor: move functions to most appropriate models
